pyproject.toml
[tool.poetry]
name = "fastapi-exercise"
version = "0.0.1"
description = "FASTAPI"
authors = ["phuccao <caohoangphuctd@gmail.com>"]
packages = [
    {include = "app"}
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0.3"
email-validator = "^2.0.0.post2"
requests = "^2.31.0"
pyyaml = "^6.0.1"
asyncio = "^3.4.3"
pydantic-settings = "^2.0.3"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pycryptodomex = "^3.19.0"
sqladmin = [
    {version="^0.16.0", extras=["full"]}
]
sqlalchemy-utils = "^0.41.1"
starlette-context = "^0.3.6"
asyncpg = "^0.29.0"
orjson = "^3.10.3"
greenlet = "^3.0.3"
alembic = "^1.13.1"
gunicorn = "^22.0.0"
httpx = "^0.27.0"
pymysql = "^1.1.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"
polyfactory = "^2.6.2"
pytest-cov = "^4.1.0"
moto = "^4.1.14"
pytest-asyncio = "^0.21.1"
flower = "^2.0.1"
autopep8 = "^2.0.4"

[tool.coverage.run]
omit = [
    # omit anything in a .local directory anywhere
    '*/.local/*',
    'tests/*',
    '__init__.py',
    '*/tests/*',
    # omit anything in a .venv directory anywhere
    '.venv/*',
]

[tool.coverage.report]
skip_empty = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"